Fixed Analyzer's methods outputing wrong results
Previously, the analyzer's methods for analyzing Danger (A*), LoS, Crazymess were completly wrong. By normalizing the difference vector and multiplying by a number from 0-1 (varying from the starting time at the parent node, and reaching the current node), when in theory the position of the sum of P2 and PD*delta should result in P1, it obviously wouldn't, since PD would be normalized, and that times 1 would result in just adding one unit in that direction, instead of resulting in P1. Therefore, all previous results are false for those metrics.

- Fixed heavy logic bugs in Analyzer LoS, Danger (A*) and Crazyness
- Fixed exception being thrown in Crazyness analysis